Handle rpm versions with x.y.z.w version numbers in test_symlinks.py
A check at the top of this file controls tests in this file.  It's
based around the version number of rpm.  The lines assumed the version
follows the x.y.z numbering scheme, but fails on anything else.  This
patch handles exceptions and just looks at the first three parts of
the version number.