Add abi.c, the code that reads in the ABI compat level files (#144)
The format is described in data/abi/GENERIC

The abidiff inspection will call read_abi() once and use that
structure for each call to the abidiff driver.  When the inspection is
done, it calls free_abi() to clean up.

This is the utility code to read in the file from the rpminspect-data
package.  The code that actually does stuff with it is in the next
commit.  That will be in the actual abidiff inspection.