No, just clone libabigail from git and build it manually on arch
The Arch User Repository won't let you build things as root and I
don't want to jump through more hoops to build a package as non-root
then install it.  We'll just clone libabigail from the git repo, find
the latest release tag, switch to that, and build it from source.