Add beginning of kmidiff inspection, put ABI functions in abi.c
The abidiff inspection is working and the bulk of its functionality
will be used by the kmidiff inspection.  This commit moves a lot of
that generic functionality over to abi.c.  The get_dirs() function has
been replaced with get_abi_dir_arg() to make it a little more
manageable.

The kmidiff inspection is not yet complete, but the framework is
there.  There is also a config file block for it in generic.yaml.

This commit also fixes #184 which was a problem with initializing the
ABI compatibility level structures.