Read list of possible kernel executable filenames from the config file.
These are stored in ri->kernel_filenames and are checked by the
kmidiff inspection (and maybe others) to determine if a file we are
looking at is a kernel exectuable file or not.  Since the kernel
filename varies by platform, let the config file specify a list of
possible names.  Default to a list containing "vmlinux" and "vmlinuz"
just so we have something to fall back on.